split plugin init package
From https://github.com/containerd/containerd/commit/70ffb12c1bed2bf9f104a4edb2256fcb363f800b
