+++ fuseoverlayfs.go
@@ -29,8 +29,6 @@ import (
 
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/containerd/mount"
-	"github.com/containerd/containerd/platforms"
-	"github.com/containerd/containerd/plugin"
 	"github.com/containerd/containerd/snapshots"
 	"github.com/containerd/containerd/snapshots/storage"
 	"github.com/containerd/continuity/fs"
@@ -40,30 +38,6 @@ import (
 const (
 	fuseoverlayfsBinary = "fuse-overlayfs"
 )
-
-func init() {
-	plugin.Register(&plugin.Registration{
-		Type:   plugin.SnapshotPlugin,
-		ID:     "fuse-overlayfs",
-		Config: &Config{},
-		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			ic.Meta.Platforms = append(ic.Meta.Platforms, platforms.DefaultSpec())
-
-			config, ok := ic.Config.(*Config)
-			if !ok {
-				return nil, errors.New("invalid fuse-overlayfs configuration")
-			}
-
-			root := ic.Root
-			if config.RootPath != "" {
-				root = config.RootPath
-			}
-
-			ic.Meta.Exports["root"] = root
-			return NewSnapshotter(root)
-		},
-	})
-}
 
 // SnapshotterConfig is used to configure the overlay snapshotter instance
 type SnapshotterConfig struct {
